refactor: rename game.ts to game config
