+++ libs/types/src/core/game/config/game-config.ts
